refactor: extract analytic shape-gradient and diagnostics from bending_tilt_leaflet.py
